backend/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
httpx>=0.26.0
mcp>=1.0.0
